feat: Phase 4 - Code structure reorganization
- Split gpu/performance_benchmark.rs (1459 lines) → gpu/benchmark/ module
  - config.rs: Benchmark configuration
  - result.rs: Result data structures
  - suite.rs: Main benchmark logic
- Split visualization/graph_viz.rs (883 lines) → visualization/graph/ module
  - nodes.rs: Graph node types
  - edges.rs: Graph edge types
  - styles.rs: Styling definitions
  - layouts.rs: Layout algorithms
  - visualizer.rs: Main visualizer
- Split visualization/tensor_viz.rs (832 lines) → visualization/tensor/ module
  - colormap.rs: Color mapping
  - config.rs: Plot configuration
  - visualizer.rs: Tensor visualization
