+++ src/gpu/benchmark/mod.rs
@@ -1,0 +1,10 @@
+//! GPU Performance Benchmark Module
+//! GPUパフォーマンスベンチマークモジュール
+
+pub mod config;
+pub mod result; 
+pub mod suite;
+
+pub use config::BenchmarkConfig;
+pub use result::BenchmarkResult;
+pub use suite::PerformanceBenchmark;
