+++ src/visualization/tensor/mod.rs
@@ -1,0 +1,10 @@
+//! Tensor Visualization Module
+//! テンソル可視化モジュール
+
+pub mod config;
+pub mod colormap;
+pub mod visualizer;
+
+pub use colormap::ColorMap;
+pub use config::TensorPlotConfig;
+pub use visualizer::TensorVisualizer;
